Atualização. Biblioteca collection.py
