+++ collection/aviao.txt
@@ -1,2 +1,2 @@
 O avião voa alto, mas o foguete voa mais.
-Mas não é qualquer um que anda de foquete.
+Mas não é qualquer um que anda de foguete.
