+++ requirements.txt
@@ -1,2 +1,5 @@
 Django==2.0.5
-psycopg2
+nltk==3.3
+psycopg2==2.7.4
+pytz==2018.4
+six==1.11.0
